requirements.txt
Flask
Pillow
websocket-client
uuid